requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
certifi==2026.2.25
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
frozenlist==1.8.0
idna==3.11
kiwisolver==1.4.9
magic-filter==1.0.12
matplotlib==3.10.8
multidict==6.7.1
numpy==2.4.1
packaging==26.0
pillow==12.1.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
speedtest-cli==2.1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
